Move repo to only contain the final complete code
We made a decision to remove the submodules in this repo for a few
reasons, but for context here is how it was setup prior to this commit:

- Each folder was it's own submodule that contained the submodule of
  each chapter which is kind of confusing.
- Each chapter was it's own branch

Because of this getting people to contribute or make changes was hard.
On top of that we have a huge amount of churn right now and expecting
the tutorial to remain the same enough to allow this is way to hard to
do. As a result we decided that we just need to only have the full
finished code from the book, rather than each individual step.